Multiuser changes - 2 June 2020
