+++ app/build.gradle
@@ -4,7 +4,9 @@ apply plugin: 'com.google.gms.google-services'
 android {
     compileSdkVersion 29
     buildToolsVersion "29.0.3"
-
+    dataBinding {
+        enabled = true
+    }
     defaultConfig {
         applicationId "com.example.connectfour"
         minSdkVersion 16
@@ -47,4 +49,7 @@ dependencies {
     implementation 'com.firebaseui:firebase-ui-database:6.2.1'
     implementation 'androidx.multidex:multidex:2.0.1'
     implementation 'com.google.android.material:material:1.1.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.9.1'
+    implementation "androidx.recyclerview:recyclerview:1.1.0"
+    implementation "androidx.recyclerview:recyclerview-selection:1.1.0-rc01"
 }
