+++ app/src/main/java/com/example/connectfour/MainActivity.java
@@ -10,7 +10,9 @@ import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.Button;
 
+import com.example.connectfour.wifi.UserListActivity;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 public class MainActivity extends AppCompatActivity {
     Button computer, friend, profile, btn_signout,btn_rules;
@@ -69,7 +71,7 @@ public class MainActivity extends AppCompatActivity {
             //BounceInterpolator interpolator=new BounceInterpolator(0.2,20);
             //bounce_anim.setInterpolator(interpolator);
             computer.startAnimation(bounce_anim);
-            startActivity(new Intent(MainActivity.this, GameActivity.class));
+            startActivity(new Intent(MainActivity.this, GameActivity.class).putExtra("type", "single"));
         }
     }
 
@@ -77,7 +79,13 @@ public class MainActivity extends AppCompatActivity {
         public void onClick(View v) {
             Animation bounce_anim= AnimationUtils.loadAnimation(MainActivity.this,R.anim.bounce_anim);
             friend.startAnimation(bounce_anim);
-            startActivity(new Intent(MainActivity.this, GameActivity.class));
+            //startActivity(new Intent(MainActivity.this, GameActivity.class));
+            FirebaseUser currentUser = FirebaseAuth.getInstance().getCurrentUser();
+            if(currentUser == null) {
+                startActivity(new Intent(MainActivity.this, loginActivity.class));
+            } else {
+                startActivity(new Intent(MainActivity.this, UserListActivity.class));
+            }
         }
     }
